refactor(backups): move all logic to the parent
